Move src/effects/effect.rs -> src/effect.rs (and delete src/effects/ dir)
